+++ core/storage/Cargo.toml
@@ -7,10 +7,31 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+bson = { version = "1.2", features = ["u2i"] }
+cfg-if = "1.0"
 # ckb-indexer = { git = "https://github.com/KaoImin/ckb-indexer", branch = "mercury" }
 ckb-types = "0.43"
+clap = "2.33"
+dashmap = "4.0"
+lazy_static = "1.4"
 log = "0.4"
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+tokio = { version = "1.10", features = ["macros", "rt-multi-thread", "sync"] }
+tokio-stream = { version = "0.1", features = ["sync"] }
 
 common = { path = "../../common" }
-protocol = { path = "../../db/protocol" }
-xsql = { path = "../../db/xsql" }
+db_protocol = { path = "../../db/protocol", package = "protocol" }
+db_xsql = { path = "../../db/xsql", package = "xsql" }
+
+[dev-dependencies]
+arc-swap = "1.3"
+criterion = { version = "0.3", features = ["async_tokio"] }
+env_logger = "0.9"
+rand = "0.7"
+
+xsql-test = { path = "../../db/xsql-test" }
+
+[[bench]]
+name = "snowflake_bench"
+harness = false
