+++ core/storage/src/relational/snowflake.rs
@@ -1,4 +1,4 @@
-use dashmap::DashMap;
+use dashmap::DashMap; 
 
 use std::sync::atomic::{AtomicI64, Ordering};
 
@@ -68,7 +68,7 @@ impl Snowflake {
 
 #[cfg(test)]
 mod tests {
-    use crate::generate_id;
+    use crate::relational::generate_id;
     use std::thread;
 
     #[test]
