+++ db/xsql/Cargo.toml
@@ -9,21 +9,12 @@ edition = "2018"
 [dependencies]
 base64 = "0.13"
 bson = { version = "1.2", features = ["u2i"] }
-cfg-if = "1.0"
 ckb-jsonrpc-types = "0.43"
 ckb-types = "0.43"
-clap = "2.33"
-dashmap = "4.0"
-futures = "0.3"
-hex = "0.4"
 is_sorted = "0.1"
-lazy_static = "1.4"
 log = "0.4"
 rbatis = { git = "https://github.com/KaoImin/rbatis.git", branch = "bson", features = ["all-database", "tokio1"] }
 serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-tokio = { version = "1.10", features = ["macros", "rt-multi-thread", "sync"] }
-tokio-stream = { version = "0.1", features = ["sync"] }
 
 common = { path = "../../common" }
 db_protocol = { path = "../protocol", package = "protocol" }
@@ -35,7 +26,3 @@ env_logger = "0.9"
 rand = "0.7"
 
 xsql-test = { path = "../xsql-test" }
-
-[[bench]]
-name = "snowflake_bench"
-harness = false
